example/demo-backend/requirements.txt
fastapi[standard]==0.115.6
pyjwt[crypto]==2.10.1
